+++ src/common/audit.js
@@ -91,9 +91,9 @@ export class Audit {
     const { Configuration } = dataAccess;
     const {
       type,
+      siteId,
       auditContext = {},
     } = message;
-    const siteId = message.url || message.siteId;
 
     try {
       const site = await this.siteProvider(siteId, context);